Books/Shopping Trips built. Authentication added.

+++ db/migrate/20200727202002_create_shopping_trips.rb
@@ -2,7 +2,7 @@ class CreateShoppingTrips < ActiveRecord::Migration[6.0]
   def change
     create_table :shopping_trips do |t|
       t.string :store
-      t.integer :receipt_total
+      t.float :receipt_total
       t.boolean :online
       t.integer :user_id
